Fix up code after rest of PR feedback
