+++ src/Abstractions/NexusMods.Abstractions.Installers/Extensions.cs
@@ -23,68 +23,7 @@ public static class Extensions
     {
         return input.ToStoredFile(to, null);
     }
-
-    /// <summary>
-    /// Convert as LibraryArchiveTree node to a LoadoutFile with the given metadata
-    /// </summary>
-    public static LoadoutFile.New ToLoadoutFile(
-        this KeyedBox<RelativePath, LibraryArchiveTree> input,
-        LoadoutId loadoutId,
-        LoadoutItemGroupId parent,
-        ITransaction tx,
-        GamePath to,
-        Optional<EntityId> entityId = default)
-    {
-        var id = entityId.ValueOr(tx.TempId);
-        var libraryFile = input.Item.LibraryFile.Value;
-
-        return new LoadoutFile.New(tx, id)
-        {
-            LoadoutItemWithTargetPath = new LoadoutItemWithTargetPath.New(tx, id)
-            {
-                LoadoutItem = new LoadoutItem.New(tx, id)
-                {
-                    LoadoutId = loadoutId,
-                    IsDisabled = false,
-                    Name = input.Item.Value.FileName,
-                    ParentId = parent,
-                },
-                TargetPath = to,
-            },
-            Hash = libraryFile.Hash,
-            Size = libraryFile.Size,
-        };
-    }
-
-    /// <summary>
-    /// Creates a group loadout item for the given library archive. Often used by installers to group files from a single archive together
-    /// </summary>
-    public static LoadoutItemGroup.New ToGroup(
-        this LibraryArchive.ReadOnly archive,
-        LoadoutId loadout,
-        ITransaction tx,
-        out LoadoutItem.New loadoutItem,
-        string? name = null,
-        Optional<EntityId> entityId = default)
-    {
-        var groupId = entityId.ValueOr(tx.TempId);
-        
-        var item = new LoadoutItem.New(tx, groupId)
-        {
-            LoadoutId = loadout,
-            IsDisabled = false,
-            Name = name ?? archive.AsLibraryFile().FileName,
-        };
-        
-        var group = new LoadoutItemGroup.New(tx, groupId)
-        {
-            LoadoutItem = item, 
-            IsGroupMarker = true,
-        };
-
-        return item;
-    }
-
+    
     /// <summary>
     /// Creates a StoredFile from a ModFileTreeSource.
     /// </summary>
