+++ src/Abstractions/NexusMods.Abstractions.Library.Installers/Extensions.cs
@@ -1,8 +1,11 @@
 using DynamicData.Kernel;
 using JetBrains.Annotations;
+using NexusMods.Abstractions.GameLocators;
 using NexusMods.Abstractions.Library.Models;
 using NexusMods.Abstractions.Loadouts;
 using NexusMods.MnemonicDB.Abstractions;
+using NexusMods.Paths;
+using NexusMods.Paths.Trees;
 
 namespace NexusMods.Abstractions.Library.Installers;
 
@@ -32,5 +35,37 @@ public static class Extensions
         };
 
         return group;
+    }
+    
+    /// <summary>
+    /// Convert as LibraryArchiveTree node to a LoadoutFile with the given metadata
+    /// </summary>
+    public static LoadoutFile.New ToLoadoutFile(
+        this KeyedBox<RelativePath, LibraryArchiveTree> input,
+        LoadoutId loadoutId,
+        LoadoutItemGroupId parent,
+        ITransaction tx,
+        GamePath to,
+        Optional<EntityId> entityId = default)
+    {
+        var id = entityId.ValueOr(tx.TempId);
+        var libraryFile = input.Item.LibraryFile.Value;
+
+        return new LoadoutFile.New(tx, id)
+        {
+            LoadoutItemWithTargetPath = new LoadoutItemWithTargetPath.New(tx, id)
+            {
+                LoadoutItem = new LoadoutItem.New(tx, id)
+                {
+                    LoadoutId = loadoutId,
+                    IsIsDisabledMarker = false,
+                    Name = input.Item.Value.FileName,
+                    ParentId = parent,
+                },
+                TargetPath = to,
+            },
+            Hash = libraryFile.Hash,
+            Size = libraryFile.Size,
+        };
     }
 }
