+++ src/NexusMods.Library/AddLibraryFileJob.cs
@@ -13,8 +13,6 @@ internal class AddLibraryFileJob : AJob
 
     public required ITransaction Transaction { get; init; }
     public required AbsolutePath FilePath { get; init; }
-    
-    public required AbsolutePath RelativeTo { get; init; }
     public required bool DoCommit { get; init; }
     public required bool DoBackup { get; init; }
