+++ src/NexusMods.Library/AddLocalFileJobWorker.cs
@@ -24,7 +24,6 @@ internal class AddLocalFileJobWorker : AJobWorker<AddLocalFileJob>
         {
             Transaction = job.Transaction,
             FilePath = job.FilePath,
-            RelativeTo = job.FilePath.Parent,
             DoCommit = false,
             DoBackup = true,
         };
