+++ tests/Games/NexusMods.Games.RedEngine.Tests/LibraryArchiveInstallerTests/PathBasedInstallerTests.cs
@@ -1,13 +1,15 @@
 using FluentAssertions;
 using NexusMods.Abstractions.GameLocators;
+using NexusMods.Games.RedEngine.Cyberpunk2077;
 using NexusMods.Games.RedEngine.ModInstallers;
+using NexusMods.Games.TestFramework;
 using NexusMods.Hashing.xxHash64;
 using NexusMods.Paths;
 using Xunit.DependencyInjection;
 
 namespace NexusMods.Games.RedEngine.Tests.LibraryArchiveInstallerTests;
 
-public class PathBasedInstallerTests : ALibraryArchiveInstallerTests
+public class PathBasedInstallerTests : ALibraryArchiveInstallerTests<Cyberpunk2077Game>
 {
     public PathBasedInstallerTests(IServiceProvider serviceProvider) : base(serviceProvider)
     {
